# internal/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "BevShop API",
    "version": "1.0.0",
    "description": "Single-owner beverage shop API. Use auth/setup once, or auth/login, then copy the returned token into Authorize (without the Bearer prefix). Request IDs must reference existing records. Dates use RFC3339 timestamps with an explicit timezone. This document describes the currently implemented routes."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/sales": {
      "get": {
        "tags": [
          "Sales"
        ],
        "summary": "List sales",
        "operationId": "sale_list",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/sale_Sale"
                  },
                  "nullable": true
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, expired or revoked token; invalid credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": false,
            "description": "Inclusive start, as an RFC3339 timestamp with timezone. Date-only strings are not supported.",
            "schema": {
              "type": "string",
              "format": "date-time",
              "example": "2026-10-01T00:00:00+02:00"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "description": "Exclusive end, as an RFC3339 timestamp with timezone. Date-only strings are not supported.",
            "schema": {
              "type": "string",
              "format": "date-time",
              "example": "2026-11-01T00:00:00+02:00"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Maximum records to return.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 50
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "description": "Number of records to skip.",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          },
          {
            "name": "payment_method",
            "in": "query",
            "required": false,
            "description": "Optional filter.",
            "schema": {
              "type": "string",
              "enum": [
                "cash",
                "mobile_money",
                "bank"
              ]
            }
          }
        ]
      },
      "post": {
        "tags": [
          "Sales"
        ],
        "summary": "Create sale",
        "operationId": "sale_create",
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sale_Sale"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body or field values.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, expired or revoked token; invalid credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Record or referenced record not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Duplicate record, record in use, or insufficient stock.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/sale_CreateRequest"
              },
              "example": {
                "payment_method": "cash",
                "notes": "",
                "items": [
                  {
                    "product_id": "11111111-1111-4111-8111-111111111111",
                    "quantity": 5,
                    "selling_price": 700
                  }
                ]
              }
            }
          }
        }
      }
    },
    "/api/v1/sales/{id}": {
      "get": {
        "tags": [
          "Sales"
        ],
        "summary": "Get sale",
        "operationId": "sale_get",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/sale_Sale"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, expired or revoked token; invalid credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Record or referenced record not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the existing record.",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "11111111-1111-4111-8111-111111111111"
          }
        ]
      }
    },
...
    "/api/v1/reports/summary": {
      "get": {
        "tags": [
          "Reports"
        ],
        "summary": "Get report summary",
        "operationId": "report_summary",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/report_Summary"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body or field values.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, expired or revoked token; invalid credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "required": false,
            "description": "Calendar period in Africa/Kigali. A valid from or to overrides this period.",
            "schema": {
              "type": "string",
              "enum": [
                "today",
                "week",
                "month",
                "year"
              ],
              "default": "today"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "description": "Inclusive start, as an RFC3339 timestamp with timezone. Date-only strings are not supported.",
            "schema": {
              "type": "string",
              "format": "date-time",
              "example": "2026-10-01T00:00:00+02:00"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "description": "Exclusive end, as an RFC3339 timestamp with timezone. Date-only strings are not supported.",
            "schema": {
              "type": "string",
              "format": "date-time",
              "example": "2026-11-01T00:00:00+02:00"
            }
          }
        ]
      }
    },
    "/api/v1/reports/dashboard": {
      "get": {
        "tags": [
          "Reports"
        ],
        "summary": "Get dashboard",
        "operationId": "report_dashboard",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/report_Dashboard"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, expired or revoked token; invalid credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": false,
            "description": "Inclusive start, as an RFC3339 timestamp with timezone. Date-only strings are not supported.",
            "schema": {
              "type": "string",
              "format": "date-time",
              "example": "2026-10-01T00:00:00+02:00"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "description": "Exclusive end, as an RFC3339 timestamp with timezone. Date-only strings are not supported.",
            "schema": {
              "type": "string",
              "format": "date-time",
              "example": "2026-11-01T00:00:00+02:00"
            }
          }
        ],
        "description": "Defaults to the current calendar month. Supply both from and to to select a custom interval."
      }
    },
    "/api/v1/history": {
      "get": {
        "tags": [
          "History"
        ],
        "summary": "Get unified history",
        "operationId": "report_history",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/report_HistoryItem"
                  },
                  "nullable": true
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, expired or revoked token; invalid credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": false,
            "description": "Inclusive start, as an RFC3339 timestamp with timezone. Date-only strings are not supported.",
            "schema": {
              "type": "string",
              "format": "date-time",
              "example": "2026-10-01T00:00:00+02:00"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "description": "Exclusive end, as an RFC3339 timestamp with timezone. Date-only strings are not supported.",
            "schema": {
              "type": "string",
              "format": "date-time",
              "example": "2026-11-01T00:00:00+02:00"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Maximum records to return.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 50
            }
          }
        ],
        "description": "Returns newest activity first. Defaults from the first day of this month to now."
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Paste the token from setup or login."
      }
    },
    "schemas": {
...
      "sale_CreateItemRequest": {
        "type": "object",
        "properties": {
          "product_id": {
            "type": "string",
            "format": "uuid",
            "example": "11111111-1111-4111-8111-111111111111"
          },
          "quantity": {
            "type": "integer",
            "example": 5,
            "minimum": 1
          },
          "selling_price": {
            "type": "number",
            "example": 700,
            "minimum": 0,
            "exclusiveMinimum": true
          }
        },
        "required": [
          "product_id",
          "quantity",
          "selling_price"
        ]
      },
      "sale_CreateRequest": {
        "type": "object",
        "properties": {
          "payment_method": {
            "type": "string",
            "example": "cash",
            "enum": [
              "cash",
              "mobile_money",
              "bank"
            ]
          },
          "notes": {
            "type": "string",
            "example": ""
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/sale_CreateItemRequest"
            },
            "nullable": false,
            "minItems": 1
          }
        },
        "required": [
          "payment_method",
          "items"
        ]
      },
      "sale_SaleItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "11111111-1111-4111-8111-111111111111"
          },
          "product_id": {
            "type": "string",
            "format": "uuid",
            "example": "11111111-1111-4111-8111-111111111111"
          },
          "product_name": {
            "type": "string",
            "example": "product_name"
          },
          "quantity": {
            "type": "integer",
            "example": 5
          },
          "selling_price_per_item": {
            "type": "number",
            "example": 0
          },
          "cost_price_per_item": {
            "type": "number",
            "example": 0
          },
          "line_total": {
            "type": "number",
            "example": 0
          },
          "line_cost": {
            "type": "number",
            "example": 0
          }
        }
      },
      "sale_Sale": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "11111111-1111-4111-8111-111111111111"
          },
          "sale_date": {
            "type": "string",
            "format": "date-time",
            "example": "2026-10-02T10:00:00+02:00"
          },
          "payment_method": {
            "type": "string",
            "example": "cash"
          },
          "total_amount": {
            "type": "number",
            "example": 0
          },
          "notes": {
            "type": "string",
            "example": ""
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/sale_SaleItem"
            },
            "nullable": true
          }
        }
      },
...
      "report_Summary": {
        "type": "object",
        "properties": {
          "sales_revenue": {
            "type": "number",
            "example": 0
          },
          "purchases": {
            "type": "number",
            "example": 0
          },
          "cost_of_items_sold": {
            "type": "number",
            "example": 0
          },
          "expenses": {
            "type": "number",
            "example": 0
          },
          "damaged_loss": {
            "type": "number",
            "example": 0
          },
          "profit_loss": {
            "type": "number",
            "example": 0
          },
          "items_sold": {
            "type": "integer",
            "example": 0
          },
          "items_purchased": {
            "type": "integer",
            "example": 0
          },
          "current_stock": {
            "type": "integer",
            "example": 0
          },
          "low_stock_count": {
            "type": "integer",
            "example": 0
          }
        }
      },
      "report_Dashboard": {
        "type": "object",
        "properties": {
          "summary": {
            "$ref": "#/components/schemas/report_Summary"
          },
          "inventory_value": {
            "type": "number",
            "example": 0
          },
          "profit_margin": {
            "type": "number",
            "example": 0
          },
          "top_products": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/report_ProductSummary"
            },
            "nullable": true
          }
        }
      },
      "report_ProductSummary": {
        "type": "object",
        "properties": {
          "product_id": {
            "type": "string",
            "format": "uuid",
            "example": "11111111-1111-4111-8111-111111111111"
          },
          "product_name": {
            "type": "string",
            "example": "product_name"
          },
          "quantity": {
            "type": "integer",
            "example": 5
          },
          "revenue": {
            "type": "number",
            "example": 0
          }
        }
      },
      "report_HistoryItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "11111111-1111-4111-8111-111111111111"
          },
          "type": {
            "type": "string",
            "example": "money_added"
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "example": "2026-10-02T10:00:00+02:00"
          },
          "description": {
            "type": "string",
            "example": ""
          },
          "amount": {
            "type": "number",
            "example": 5000
          },
          "reference_id": {
            "type": "string",
            "format": "uuid",
            "example": "11111111-1111-4111-8111-111111111111"
          }
        }
      },
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "example": "invalid request body"
          }
        }
      },
...
    }
  }
}